UltronConfig params
