+++ ultron-allure/src/main/java/com/atiurin/ultron/allure/config/AllureAttachStrategy.kt
@@ -1,0 +1,9 @@
+package com.atiurin.ultron.allure.config
+
+enum class AllureAttachStrategy {
+    TEST_FAILURE,
+    OPERATION_FAILURE, // attach artifact for failed operation
+    OPERATION_SUCCESS, // attach artifact for each succeeded operation
+    OPERATION_FINISH, // attach artifact for each operation
+    NONE
+}
