+++ ultron-allure/src/main/java/com/atiurin/ultron/allure/config/AllureConfigParams.kt
@@ -1,0 +1,11 @@
+package com.atiurin.ultron.allure.config
+
+data class AllureConfigParams(
+    var addScreenshotPolicy: MutableSet<AllureAttachStrategy> = mutableSetOf(
+        AllureAttachStrategy.TEST_FAILURE,
+        AllureAttachStrategy.OPERATION_FAILURE
+    ),
+    var attachUltronLog: Boolean = true,
+    var addConditionsToReport: Boolean = true,
+    var detailedAllureReport: Boolean = true
+)
