+++ ultron/src/main/java/com/atiurin/ultron/runner/UltronLogRunListener.kt
@@ -2,7 +2,6 @@ package com.atiurin.ultron.runner
 
 import com.atiurin.ultron.extensions.fullTestName
 import com.atiurin.ultron.log.LogLevel
-import com.atiurin.ultron.log.UltronLog
 import com.atiurin.ultron.log.UltronLogUtil.logTextBlock
 import org.junit.runner.Description
 import org.junit.runner.Result
@@ -51,6 +50,4 @@ class UltronLogRunListener : UltronRunListener() {
             """.trimMargin()
         )
     }
-
-
 }
